Remove std::copy & size_t to int cast in ByteBuf

+++ common/src/byteBuffer.cpp
@@ -1,7 +1,7 @@
 #include "byteBuffer.h"
 
 
-int ByteBuffer::Length() const
+size_t ByteBuffer::Length() const
 {
     return internal.size();
 }
